[s2s trainer] tests to use distributed on multi-gpu machine (#7965)
